@@ -9,7 +9,16 @@ exports.processESLintMessages = exports.generateDebugString = exports.getDebugIn
 let getDebugInfo = exports.getDebugInfo = (() => {
   var _ref = _asyncToGenerator(function* (worker) {
     const textEditor = atom.workspace.getActiveTextEditor();
-    const filePath = textEditor.getPath();
+    let filePath;
+    let editorScopes;
+    if (atom.workspace.isTextEditor(textEditor)) {
+      filePath = textEditor.getPath();
+      editorScopes = textEditor.getLastCursor().getScopeDescriptor().getScopesArray();
+    } else {
+      // Somehow this can be called with no active TextEditor, impossible I know...
+      filePath = 'unknown';
+      editorScopes = ['unknown'];
+    }
     const packagePath = atom.packages.resolvePackagePath('linter-eslint');
     let linterEslintMeta;
     if (packagePath === undefined) {
@@ -37,7 +46,8 @@ let getDebugInfo = exports.getDebugInfo = (() => {
         hoursSinceRestart,
         platform: process.platform,
         eslintType: response.type,
-        eslintPath: response.path
+        eslintPath: response.path,
+        editorScopes
       };
     } catch (error) {
       atom.notifications.addError(`${error}`);
@@ -53,7 +63,7 @@ let getDebugInfo = exports.getDebugInfo = (() => {
 let generateDebugString = exports.generateDebugString = (() => {
   var _ref2 = _asyncToGenerator(function* (worker) {
     const debug = yield getDebugInfo(worker);
-    const details = [`Atom version: ${debug.atomVersion}`, `linter-eslint version: ${debug.linterEslintVersion}`, `ESLint version: ${debug.eslintVersion}`, `Hours since last Atom restart: ${debug.hoursSinceRestart}`, `Platform: ${debug.platform}`, `Using ${debug.eslintType} ESLint from ${debug.eslintPath}`, `linter-eslint configuration: ${JSON.stringify(debug.linterEslintConfig, null, 2)}`];
+    const details = [`Atom version: ${debug.atomVersion}`, `linter-eslint version: ${debug.linterEslintVersion}`, `ESLint version: ${debug.eslintVersion}`, `Hours since last Atom restart: ${debug.hoursSinceRestart}`, `Platform: ${debug.platform}`, `Using ${debug.eslintType} ESLint from: ${debug.eslintPath}`, `Current file's scopes: ${JSON.stringify(debug.editorScopes, null, 2)}`, `linter-eslint configuration: ${JSON.stringify(debug.linterEslintConfig, null, 2)}`];
     return details.join('\n');
   });
 
