@@ -176,7 +176,9 @@ function getArgv(type, config, rules, filePath, fileDir, givenConfigPath) {
     if (!_path2.default.isAbsolute(rulesDir)) {
       rulesDir = (0, _atomLinter.findCached)(fileDir, rulesDir);
     }
-    argv.push('--rulesdir', rulesDir);
+    if (rulesDir) {
+      argv.push('--rulesdir', rulesDir);
+    }
   }
   if (configPath) {
     argv.push('--config', (0, _resolveEnv2.default)(configPath));
